finished refactoring the endpoints
